+++ src/styles/resolvers/position/index.ts
@@ -1,0 +1,12 @@
+import { responsiveVariables } from "utilities/props";
+import * as T from "styles/types";
+import "./position.css";
+
+const position: T.StyleResolver<T.Position> = (value) => {
+	if (!value) return {};
+	const variables = responsiveVariables("--rs-position", value);
+
+	return { variables };
+};
+
+export default position;